In --paired mode, the "group" tool now reports the position of the 2nd mate
The output contains a new column "position2" which contains the position of
the 2nd read, computed in the same way as the position of read1 (i.e, the
position is the position of the 1st base in sequencing (not reference!)
direction, and is adjusted for possible clipped bases.

This uses the previously introduced support for collecting read pairs in
get_bundles.