Made get_bundles use get_readpairs to process mate pairs together
In --paired mode, get_bundles now uses get_readpairs to process the first
and second read of a pair together. The 2nd reads are now stored in the bundle
as well, in a list parallel to "read" called "read2". This currently works
only for all_reads=True.

So far, the second read is only stored. The next commits will introduce
changes that use the information from the second read to improve the grouping.