Fix travis-ci osx builds.
Apparently, python3 is now pre-installed, and attempting to install
python2 using homebrew thus fails. So I changed .travis.yml to do
"brew upgrade python", since that is what the previous error message
suggested doing, and made sure that the installation directory is on
the current PATH. After upgrading to python3, virtualenv seems to be
broken -- it complains about being unable to find python2.7. So I
addded a pip command to re-install virtualenv for python3.